requirements.txt
streamlit
streamlit-authenticator
pyyaml
langchain
langchain-community
langchain-core
langchain-ollama
ollama
bcrypt
